[refactor] 부스 상세 화면 현재 운영 중인 부스 판단 로직 Util로 이동
